UI for registering department
Created the JFrame for the register department page, and implemented all
of its functionality, including error checking. The LogIn page will now
no longer allow a department to be created, as one exists in the output
xml file. This is by design. Kind of a lame way to get past umple
getting confused by setting a child class as a singleton, but I digress.
It works. Signing in with the ID 260690005 will sign you in as the
department (still won't send you to a new page though)... in fact, the
application currently never signs out, so... the department is "signed
in" according to the model whenever you open it up. Signing out will be
implemented later to occur whenever a "close" operation occurs (on a
"signed in" view), or whenever a log out button is pushed.