Post Job Page UI nearing completion
The post job page is looking much better than before, but some sizing
could be improved. When grader job is selected, min hours and max hours
input components are replaced by an hours input component, and the lab
session check box is hidden. The functionality for posting a job has not
yet been implemented, however the post job page can be accessed by the
department and any instructor. The course list in the drop down box
should show every course if the department is logged in, and only the
instructor's courses when an instructor is logged in. Also, the back
button sends the instructor back to the instructor's main page, and the
department gets sent back to the department's main page. Both of these
main pages currently look the same, except for the title. In any "logged
in" view, pressing the exit button logs the user out, and exits the
program.