Course Management View complete
The department can now add courses, and add instructors to courses (up
to 4 instructors per course). I believe this also means that TA jobs can
be added as well (although I haven't tested this). I made a few minor
changes to the model, like adding the number of students for a course,
and switching contact times to be associated with Instructors instead of
courses. The  Instructor management view (will be accessible to
Instructors and the department) was also begun.