Job applications by applicants and the department
Job applications can now be made with presumably complete error
checking. The department can make applications on behalf of any
applicant, and applicants can make applications as well. This concludes
the work for deliverable 2. Ignore the errors in the project, as these
appear in one view class, and this class is not yet used in the app.