More error checking
Added an error check for applying to more than 3 jobs, and also fixed a
bug in the error checking for posting TA jobs.
Some naming needs to be changed, and the deadlines for posting a job and
applying to a job should be visible in the corresponding views. That
does not need to be implemented for this deliverable.