Model Changes
The model had a few missing attributes. This update fixed that, and also
made the necessary updates to the controller and the views to make the
app workable once more.
Model changes:
Job now has a Boolean isApproved.
Course now has an Integer Budget.
Application now has a Boolean isAccepted.