Application Ranker
A controller method was implemented which splits a job's list of
applicants into 4 different lists (tut and Lab graduates, tut and lab
undergraduates, graduates, and undergraduates), and returns a list of
these four lists in order of priority. These 4 lists will be chosen from
for display in a comboBox by a selection of radiobuttons. These buttons
will act as a "filter" for the department to view applicants based on
the designated priority mentioned in the User Story.