Job Posting Update (Android Implications)
Changed the job posting process to automatically approve jobs that are
posted by the department. The controller methods for posting jobs were
changed to include another boolean input (approval). The corresponding
views' calls to these methods were changed to reflect this, and they
must be changed in the Android App as well.