Some view edits, and AssignJobsPage
Some text areas were not extending down when the text was too large for
the area, so these were added to scrollPanes. Some of them, at least.
Eclipse crashed multiple times when I tried to edit some of the pages to
put the text areas in scroll panes, so I decided to leave it as is, and
hope that they extend downwards properly.
The AssignJobsPage was completed.